Support signing a pre-calculated hash (#87)
* Split the hashing out of the sign method

This code change adds support to split the hashing of a message
and the actual signing of the message.

* Updating unit test and documentation

This commit updates the unit test and usage docs. In addition,
This change removes a redundant error check inside rsa.sign().

* Refactore unit tests and code comments

Removed the print statements from the unit test and refactored a
few code comments to improve readability.

* Rename hash function

The new hash function had the same name as a function in the
standard library. This commit changes the name to avoid conflicts.

* Rename hash function to compute_hash()

This commit renames the hash function to compute_hash().